requirements.txt
Flask==3.0.3
psycopg==3.1.18
pytest==8.1.1
python-dotenv==1.0.1
